exp_backend/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
web3
